Consolidate driver code into llvm-stress.cpp
